feat: 重構配置模組以支持新環境
- 將環境配置模組重構為 Façade，保持向後相容性，並將實際定義移至子模組中。
- 新增多個環境檢測與配置相關的模組，提升代碼可維護性與可讀性。
- 整合安全驗證常量與時序配置，簡化存儲鍵的管理。
- 引入新的建置環境配置範本，支持 CI/CD 流程中的動態生成與注入 Secrets。
