+++ scripts/config/shared/core/index.js
@@ -1,0 +1,7 @@
+/**
+ * Core 配置聚合入口
+ */
+
+export * from './browser.js';
+export * from './timing.js';
+export * from './security.js';
