+++ scripts/config/shared/core/security.js
@@ -1,0 +1,64 @@
+/**
+ * 安全驗證常量
+ * SVG 白名單、安全 URL 協議等
+ */
+
+export const SECURITY_CONSTANTS = {
+  SVG_ALLOWED_TAGS: [
+    'svg',
+    'path',
+    'circle',
+    'rect',
+    'line',
+    'polyline',
+    'polygon',
+    'ellipse',
+    'g',
+    'defs',
+    'use',
+    'symbol',
+    'title',
+    'desc',
+    'lineargradient',
+    'radialgradient',
+    'stop',
+    'clippath',
+    'mask',
+    'pattern',
+    'text',
+    'tspan',
+    'image',
+    'a',
+  ],
+
+  SVG_ALLOWED_ATTRS: [
+    'viewBox',
+    'fill',
+    'stroke',
+    'stroke-width',
+    'stroke-linecap',
+    'stroke-linejoin',
+    'width',
+    'height',
+    'class',
+    'id',
+    'xmlns',
+    'x',
+    'y',
+    'cx',
+    'cy',
+    'r',
+    'rx',
+    'ry',
+    'd',
+    'points',
+    'x1',
+    'y1',
+    'x2',
+    'y2',
+    'transform',
+    'opacity',
+  ],
+
+  SAFE_URL_PROTOCOLS: ['http:', 'https:', ''],
+};
