+++ scripts/config/shared/core/timing.js
@@ -1,0 +1,18 @@
+/**
+ * 時序配置
+ * Handler 重試延遲、TabService 超時等時間相關常數
+ */
+
+export const HANDLER_CONSTANTS = {
+  BUNDLE_READY_RETRY_DELAY: 150,
+  BUNDLE_READY_MAX_RETRIES: 10,
+  PAGE_STATUS_CACHE_TTL: 60 * 1000,
+  IMAGE_RETRY_DELAY: 500,
+  CHECK_DELAY: 500,
+};
+
+export const TAB_SERVICE = {
+  LOADING_TIMEOUT_MS: 10_000,
+  STATUS_UPDATE_DELAY_MS: 1000,
+  PRELOADER_PING_TIMEOUT_MS: 500,
+};
