+++ scripts/config/shared/storage/auth.js
@@ -1,0 +1,13 @@
+/**
+ * 認證相關 Storage Keys
+ * 存放於 chrome.storage.local
+ */
+
+export const AUTH_LOCAL_KEYS = [
+  'notionAuthMode',
+  'notionOAuthToken',
+  'notionRefreshToken',
+  'notionRefreshProof',
+  'notionWorkspaceName',
+  'notionAuthEpoch',
+];
