+++ scripts/config/shared/storage/index.js
@@ -1,0 +1,8 @@
+/**
+ * Storage 配置聚合入口
+ */
+
+export * from './keys.js';
+export * from './prefixes.js';
+export * from './auth.js';
+export * from './dataSource.js';
