+++ scripts/config/shared/storage/keys.js
@@ -1,0 +1,15 @@
+/**
+ * Sync 設定鍵
+ * 跨裝置共用的使用者設定，存放於 chrome.storage.sync
+ */
+
+export const SYNC_CONFIG_KEYS = [
+  'notionApiKey',
+  'titleTemplate',
+  'addSource',
+  'addTimestamp',
+  'highlightStyle',
+  'highlightContentStyle',
+  'enableDebugLogs',
+  'uiZoomLevel',
+];
