+++ scripts/config/shared/storage/prefixes.js
@@ -1,0 +1,9 @@
+/**
+ * Storage 前綴常量
+ */
+
+export const SAVED_PREFIX = 'saved_';
+export const HIGHLIGHTS_PREFIX = 'highlights_';
+export const URL_ALIAS_PREFIX = 'url_alias:';
+export const HIGHLIGHTS_LEGACY_PREFIX = 'nh_highlights_';
+export const PAGE_PREFIX = 'page_';
